requirements.txt
Python==3.8
newsapi-python==0.2.6

